beta-1.1.1
+ make minecart experiment on by default (old worlds will have it automatically added on update)
+ add inferno rails functionality
+ add 3 new advancements relating to rails
+ allow inferno rails to grant "wax on" & "wax off" advancements
+ modify all rail recipes
+ add more minecart recipes

+++ common/src/main/java/sircow/preservedinferno/mixin/MagmaCubeMixin.java
@@ -18,12 +18,8 @@ public class MagmaCubeMixin {
         this.sizeTemp = size;
     }
 
-    @ModifyArg(
-            method = "setSize(IZ)V",
-            at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/ai/attributes/AttributeInstance;setBaseValue(D)V"),
-            index = 0
-    )
+    @ModifyArg(method = "setSize(IZ)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/ai/attributes/AttributeInstance;setBaseValue(D)V"), index = 0)
     private double preserved_inferno$modifyArmorValue(double baseValue) {
-        return (double)(this.sizeTemp * 15);
+        return this.sizeTemp * 15;
     }
 }
